Clean up top-level model component
